2022.08.11 jo

+++ team_mypet/src/main/java/com/jeonju/mypet/controller/SellerController.java
@@ -43,8 +43,10 @@ public class SellerController {
 			Model model, HttpServletRequest request) {
 		
 		HttpSession session = request.getSession();
-		int midx = (int) session.getAttribute("midx");
-		String member_id= Integer.toString(midx);
+//		int midx = (int) session.getAttribute("midx");
+		long midx = (long) session.getAttribute("midx");
+		String member_id= Integer.toString((int) midx);
+//		String member_id= Integer.toString(midx);
 		
 		if(searching == null)
 			searching="searchTotal";
@@ -142,8 +144,9 @@ public class SellerController {
 			) throws IllegalStateException, IOException{
 		
 		HttpSession session = request.getSession();
-		int midx = (int) session.getAttribute("midx");
-		String member_id= Integer.toString(midx);
+//		int midx = (int) session.getAttribute("midx");
+		long midx = (long) session.getAttribute("midx");
+		String member_id= Integer.toString((int) midx);
 		
 		String p_category_idx = "";
 		int p_category_large =  Integer.parseInt(param.get("p_category_large"));
